// src/Akeneo/Pim/Automation/FranklinInsights/Infrastructure/Symfony/Resources/public/react/domain/model/attribute-type.enum.ts
/*
 * This file is part of the Akeneo PIM Enterprise Edition.
 *
 * (c) 2019 Akeneo SAS (http://www.akeneo.com)
 *
 * For the full copyright and license information, please view the LICENSE
 * file that was distributed with this source code.
 */

export enum AttributeType {
  BOOLEAN = 'pim_catalog_boolean',
  METRIC = 'pim_catalog_metric',
  MULTISELECT = 'pim_catalog_multiselect',
  NUMBER = 'pim_catalog_number',
  SIMPLESELECT = 'pim_catalog_simpleselect',
  TEXT = 'pim_catalog_text',
  TEXTAREA = 'pim_catalog_textarea'
}